Use uint8_t and int8_t types

+++ touchpad-piano/MIDITouchpadPiano.hpp
@@ -12,7 +12,7 @@ BEGIN_CS_NAMESPACE
 #define HOLD_THRESHOLD_MS 500
 
 namespace PianoMode {
-    enum Mode: byte {Standard, Hold, Monodic};
+    enum Mode: uint8_t {Standard, Hold, Monodic};
 } // namespace PianoMode
 
 
@@ -32,8 +32,8 @@ public:
         // Too high rate reading is no working
         if (!updateTimer) return;
 
-        byte newKeysState[NB_NOTES];
-        int newMonodicNote = -1;
+        uint8_t newKeysState[NB_NOTES];
+        int8_t newMonodicNote = -1;
 
         readData(newKeysState, &newMonodicNote);
         
@@ -56,7 +56,7 @@ public:
         // Leaving Hold mode requires to stop notes
         if (this->mode == PianoMode::Hold)
         {
-            for (byte i = 0; i < NB_NOTES; i++)
+            for (uint8_t i = 0; i < NB_NOTES; i++)
             {
                 if (keysState[i] == 1)
                 {
@@ -70,9 +70,9 @@ public:
 
     PianoMode::Mode getMode() { return mode; }
 
-    void readData(byte newKeysState[NB_NOTES], int* newMonodicNote)
+    void readData(uint8_t newKeysState[NB_NOTES], int8_t* newMonodicNote)
     {
-        for (byte i = 0; i < NB_NOTES; i++)
+        for (uint8_t i = 0; i < NB_NOTES; i++)
         {
             digitalWrite(sclPin, LOW);
             newKeysState[NoteMap[i]] = !digitalRead(sdaPin);
@@ -86,9 +86,9 @@ public:
         }
     }
 
-    void handleStandard(byte newKeysState[NB_NOTES])
+    void handleStandard(uint8_t newKeysState[NB_NOTES])
     {
-        for (byte i = 0; i < NB_NOTES; i++)
+        for (uint8_t i = 0; i < NB_NOTES; i++)
         {
             // Key changed state
             if (newKeysState[i] != keysState[i])
@@ -113,11 +113,11 @@ public:
 #endif
     }
 
-    void handleHold(byte newKeysState[NB_NOTES])
+    void handleHold(uint8_t newKeysState[NB_NOTES])
     {
         uint64_t holdPrevTouch = holdLastTouchMs;
 
-        for (byte i = 0; i < NB_NOTES; i++)
+        for (uint8_t i = 0; i < NB_NOTES; i++)
         {
             if ((newKeysState[i] & keysState[i]) == 1)
             {
@@ -155,7 +155,7 @@ public:
 #endif
     }
 
-    void handleMonodic(int newMonodicNote)
+    void handleMonodic(int8_t newMonodicNote)
     {
         if (newMonodicNote != monodicNote)
         {
@@ -173,12 +173,12 @@ private:
     const pin_t sclPin;
     const pin_t sdaPin;
     const MIDIAddress baseAddress;
-    const byte NoteMap[NB_NOTES] = {1,3,6,8,10,0,2,4,5,7,9,11};
-    byte keysState[NB_NOTES] = {0};
+    const uint8_t NoteMap[NB_NOTES] = {1,3,6,8,10,0,2,4,5,7,9,11};
+    uint8_t keysState[NB_NOTES] = {0};
     uint64_t holdLastTouchMs = 0;
     AH::Timer<millis> updateTimer;
     PianoMode::Mode mode;
-    int monodicNote;
+    int8_t monodicNote;
 
 public:
     DigitalNoteSender sender;
